fix: fetch snipept by id issues
