+++ client/src/_context/code/code.type.ts
@@ -21,6 +21,7 @@ export interface State {
     updateInput: (input: string) => void;
     updateLanguage: (lang: string) => void;
     submitCode: (code: string, language: string, input: string) => void;
+    fetchSnippetbyId: (snipId: string) => void;
 }
 
 export interface Action {
