+++ client/src/_context/room/room.reducer.ts
@@ -5,7 +5,7 @@ import {
     UPDATE_ROOM_CODE,
     UPDATE_ROOM_INPUT,
     UPDATE_ROOM_LANGUAGE,
-    UPDATE_ROOM_OUTPUT
+    UPDATE_ROOM_OUTPUT,
 } from '_context/types';
 import { Action, State } from './room.type';
